+++ library/src/schemas/object/objectAsync.ts
@@ -11,7 +11,7 @@ import type { ObjectInput, ObjectOutput } from './types.ts';
 /**
  * Object shape async type.
  */
-export type ObjectShapesAsync = Record<
+export type ObjectShapeAsync = Record<
   string,
   BaseSchema<any> | BaseSchemaAsync<any>
 >;
@@ -20,7 +20,7 @@ export type ObjectShapesAsync = Record<
  * Object schema async type.
  */
 export type ObjectSchemaAsync<
-  TObjectShape extends ObjectShapesAsync,
+  TObjectShape extends ObjectShapeAsync,
   TOutput = ObjectOutput<TObjectShape>
 > = BaseSchemaAsync<ObjectInput<TObjectShape>, TOutput> & {
   schema: 'object';
@@ -35,7 +35,7 @@ export type ObjectSchemaAsync<
  *
  * @returns An async object schema.
  */
-export function objectAsync<TObjectShape extends ObjectShapesAsync>(
+export function objectAsync<TObjectShape extends ObjectShapeAsync>(
   object: TObjectShape,
   pipe?: PipeAsync<ObjectOutput<TObjectShape>>
 ): ObjectSchemaAsync<TObjectShape>;
@@ -49,13 +49,13 @@ export function objectAsync<TObjectShape extends ObjectShapesAsync>(
  *
  * @returns An async object schema.
  */
-export function objectAsync<TObjectShape extends ObjectShapesAsync>(
+export function objectAsync<TObjectShape extends ObjectShapeAsync>(
   object: TObjectShape,
   error?: string,
   pipe?: PipeAsync<ObjectOutput<TObjectShape>>
 ): ObjectSchemaAsync<TObjectShape>;
 
-export function objectAsync<TObjectShape extends ObjectShapesAsync>(
+export function objectAsync<TObjectShape extends ObjectShapeAsync>(
   object: TObjectShape,
   arg2?: PipeAsync<ObjectOutput<TObjectShape>> | string,
   arg3?: PipeAsync<ObjectOutput<TObjectShape>>