+++ library/src/types.ts
@@ -75,3 +75,19 @@ export type PipeAsync<TValue> = ((
   value: TValue,
   info: ValidateInfo
 ) => TValue | Promise<TValue>)[];
+
+/**
+ * Resolve type.
+ *
+ * Hint: This type has no effect and is only used so that TypeScript displays
+ * the final type in the preview instead of the utility types used.
+ */
+type Resolve<T> = T;
+
+/**
+ * Resolve object type.
+ *
+ * Hint: This type has no effect and is only used so that TypeScript displays
+ * the final type in the preview instead of the utility types used.
+ */
+export type ResolveObject<T> = Resolve<{ [k in keyof T]: T[k] }>;