+++ library/CHANGELOG.md
@@ -2,11 +2,16 @@
 
 All notable changes to the library will be documented in this file.
 
+## v0.8.0 (July 31, 2023)
+
+- Fix infered `object` and `record` types (issue #9, #10, #34)
+- Add `strict` and `strictAsync` method to detect unknown object keys
+
 ## v0.7.0 (July 30, 2023)
 
 - Add `is` method which can be used as a type guard (pull request #13)
-- Throw all validation issues of a pipeline by default (issues #18)
-- Add `abortPipeEarly` option to abort pipe on first error (issues #18)
+- Throw all validation issues of a pipeline by default (issue #18)
+- Add `abortPipeEarly` option to abort pipe on first error (issue #18)
 - Add `abortEarly` option to abort on first error
 
 ## v0.6.0 (July 30, 2023)